bug: [ON-2864] handle missing data from globalAPI &  rename module
